Split out part of RunningSummarizer into SummaryGenerator
